refactor: ui model 분리
