+++ app/src/main/java/com/bestapp/zipbab/ui/meetupmap/model/MeetUpMapUiState.kt
@@ -1,4 +1,4 @@
-package com.bestapp.zipbab.ui.meetupmap
+package com.bestapp.zipbab.ui.meetupmap.model
 
 data class MeetUpMapUiState(
     val meetUpMapMeetingUis: List<MeetUpMapUi> = emptyList(),
